+++ examples/counter_cubit/lib/counter_cubit.dart
@@ -1,19 +1,9 @@
 import 'package:bloc/bloc.dart';
-import 'package:bloc_annotation/bloc_annotation.dart';
+import 'package:bloc_meta/bloc_meta.dart';
 
 part 'counter_cubit.g.dart';
 
-// @CubitMeta(state: int)
-// final class CounterCubit extends _$CounterCubit {
-//   CounterCubit() : super(0);
-//
-//   void increment() => emit(state + 1);
-//
-//   void decrement() => emit(state - 1);
-//   void reset() => emit(0);
-// }
-
-@CubitClass<int>()
+@CubitMeta<int>()
 final class CounterCubit extends _$CounterCubit {
   CounterCubit() : super(0);