+++ packages/bloc_meta_generator/lib/src/types.dart
@@ -1,0 +1,34 @@
+/// Properties described in [BaseAnnotation] from `bloc_meta` package
+typedef BaseAnnotationProperties = ({
+  String name,
+  bool copyWith,
+  bool overrideEquality,
+  bool overrideToString,
+});
+
+/// Properties described in [Cubit] annotaion from `bloc_meta` package
+typedef CubitAnnotationProperties = ({
+  String name,
+  bool copyWith,
+  bool overrideEquality,
+  bool overrideToString,
+});
+
+/// Properties described in [Bloc] annotaion from `bloc_meta` package
+typedef BlocAnnotationProperties = ({
+  String name,
+  bool copyWith,
+  bool overrideEquality,
+  bool overrideToString,
+});
+
+/// Record type for describing a attributes of a Class
+///
+/// ```dart
+/// class Some {
+///   final String attribute1;
+///   final int attribute2;
+///   final double attribute3;
+/// }
+/// ```
+typedef ElementAttribute = ({String type, String name});