+++ packages/bloc_meta_generator/test/src/test_cubit.dart
@@ -1,0 +1,34 @@
+import 'package:bloc_meta/bloc_meta.dart';
+import 'package:source_gen_test/annotations.dart';
+
+abstract class Cubit<State> {
+  Cubit(State initialState) : state = initialState;
+  State state;
+  void emit(State state) {}
+}
+
+@ShouldGenerate(r'''
+abstract class _$TestCubit extends Cubit<int> {
+  _$TestCubit(super.initialState);
+
+  @override
+  String toString() => 'TestCubit(state: $state)';
+
+  @override
+  int get hashCode => state.hashCode;
+
+  @override
+  bool operator ==(Object other) {
+    if (identical(this, other)) return true;
+    return other is _$TestCubit && other.state == state;
+  }
+}
+''')
+@CubitMeta<int>()
+final class TestCubit extends _$TestCubit {
+  TestCubit() : super(0);
+}
+
+abstract class _$TestCubit extends Cubit<int> {
+  _$TestCubit(super.initialState);
+}